improve return value handling
* ability to call a function returned by an expression evaluation.
* ability to write to an array or object returned by an expression
  evaluation.
* fix memory leak

+++ src/je_util.c
@@ -50,7 +50,7 @@ char* je_utf8str(uint32_t c) {
 char* je_astri64(int64_t src) {
     char* dest = calloc(1, IBUFSIZE);
 
-    snprintf(dest, IBUFSIZE, "%lld", (unsigned long long) src);
+    snprintf(dest, IBUFSIZE, "%lld", src);
 
     return dest;
 }
